pyproject.toml
[project]
name = "elearning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels[daphne]>=4.3.2",
    "channels-redis>=4.3.0",
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.1",
    "celery>=5.6.2",
    "drf-yasg>=1.21.15",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
]
